+++ src-ui/app/models/sleep-mode.ts
@@ -1,9 +1,10 @@
 import { AutomationType } from './automations';
 
-export type SleepModeStatusChangeReasonType = 'MANUAL' | 'OSC_CONTROL' | 'AUTOMATION';
+export type SleepModeStatusChangeReasonType = 'MANUAL' | 'HOTKEY' | 'OSC_CONTROL' | 'AUTOMATION';
 
 export type SleepModeStatusChangeReason =
   | ManualSleepModeStatusChangeReason
+  | HotkeySleepModeStatusChangeReason
   | OSCControlSleepModeStatusChangeReason
   | AutomationSleepModeStatusChangeReason;
 
@@ -14,6 +15,10 @@ export interface SleepModeStatusChangeReasonBase {
 
 export interface ManualSleepModeStatusChangeReason extends SleepModeStatusChangeReasonBase {
   type: 'MANUAL';
+}
+
+export interface HotkeySleepModeStatusChangeReason extends SleepModeStatusChangeReasonBase {
+  type: 'HOTKEY';
 }
 
 export interface OSCControlSleepModeStatusChangeReason extends SleepModeStatusChangeReasonBase {