+++ src-ui/app/views/dashboard-view/views/sleep-detection-view/sleep-detection-view.component.ts
@@ -54,7 +54,7 @@ export class SleepDetectionViewComponent implements OnInit {
   onClick(event: MouseEvent) {
     if ((event.target as HTMLElement).className !== 'integrationsPageLink') return;
     event.preventDefault();
-    this.router.navigate(['/dashboard/settings'], { fragment: 'INTEGRATIONS' });
+    this.router.navigate(['/dashboard/settings/integrations']);
   }
 
   openModal_EnableSleepModeAtTime() {